+++ sdk/src/androidTest/java/ly/count/android/sdk/ModuleUserProfileTests.java
@@ -1,5 +1,6 @@
 package ly.count.android.sdk;
 
+import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import java.util.HashMap;
 import java.util.Map;
@@ -436,5 +437,118 @@ public class ModuleUserProfileTests {
         Assert.assertTrue(req.contains("gender"));
         Assert.assertTrue(req.contains("custom"));
         Assert.assertTrue(req.contains("byear"));
+    }
+
+    /**
+     * Test that custom data keys are truncated to the maximum allowed length (10)
+     * Due to truncation, the keys "hair_color_id" and "hair_color_tone" will be merged into "hair_color"
+     * The value of "hair_color" will be the value of "hair_color_tone" since it was set last
+     * The value of "hair_skin_tone" will be truncated to "hair_skin_"
+     * Tha last value of "hair_color" will be "black"
+     * And predefined key "picturePath" is not truncated
+     */
+    @Test
+    public void internalLimit_testCustomData() {
+        Countly mCountly = Countly.sharedInstance();
+        CountlyConfig config = new CountlyConfig(TestUtils.getContext(), "appkey", "http://test.count.ly").setDeviceId("1234").setLoggingEnabled(true).enableCrashReporting();
+        config.sdkInternalLimits.setMaxKeyLength(10);
+        mCountly.init(config);
+
+        HashMap<String, Object> data = new HashMap<>();
+        data.put("hair_color_id", 4567);
+        data.put("hair_color_tone", "bold");
+        mCountly.userProfile().setProperties(data);
+        assertEquals(1, mCountly.moduleUserProfile.custom.size());
+        assertEquals("bold", mCountly.moduleUserProfile.custom.get("hair_color"));
+
+        mCountly.userProfile().setProperty("hair_color", "black");
+        mCountly.userProfile().setProperty("hair_skin_tone", "yellow");
+        mCountly.userProfile().setProperty("picturePath", "Test Test");
+        assertEquals(2, mCountly.moduleUserProfile.custom.size());
+        assertNull(ModuleUserProfile.picturePath);
+        assertEquals("black", mCountly.moduleUserProfile.custom.get("hair_color"));
+        assertEquals("yellow", mCountly.moduleUserProfile.custom.get("hair_skin_"));
+    }
+
+    /**
+     * Test that custom data keys are truncated to the maximum allowed length (10)
+     * Due to truncation, for push keys, the keys "reminder" and "rock" will be merged into same key
+     */
+    @Test
+    public void internalLimit_testCustomModifiers() throws JSONException {
+        Countly mCountly = Countly.sharedInstance();
+        CountlyConfig config = new CountlyConfig(TestUtils.getContext(), "appkey", "http://test.count.ly").setDeviceId("1234").setLoggingEnabled(true).enableCrashReporting();
+        config.sdkInternalLimits.setMaxKeyLength(10);
+        mCountly.init(config);
+
+        mCountly.moduleUserProfile.modifyCustomData("key_inc_with", 1, "$inc");
+        mCountly.moduleUserProfile.modifyCustomData("key_mul_width", 2, "$mul");
+        mCountly.userProfile().push("key_push_reminder", "test1");
+        mCountly.userProfile().push("key_push_rock", "test3");
+
+        assertEquals(1, mCountly.moduleUserProfile.customMods.get("key_inc_wi").getInt("$inc"));
+        assertEquals(2, mCountly.moduleUserProfile.customMods.get("key_mul_wi").getInt("$mul"));
+        assertEquals(2, mCountly.moduleUserProfile.customMods.get("key_push_r").getJSONArray("$push").length());
+        assertEquals("test1", mCountly.moduleUserProfile.customMods.get("key_push_r").getJSONArray("$push").getString(0));
+        assertEquals("test3", mCountly.moduleUserProfile.customMods.get("key_push_r").getJSONArray("$push").getString(1));
+    }
+
+    /**
+     * "setProperties" with both custom and predefined properties
+     * custom properties should be truncated but predefined properties should not be truncated
+     * validate that the predefined properties are not truncated
+     */
+    @Test
+    public void internalLimit_setProperties() throws JSONException {
+        Countly mCountly = Countly.sharedInstance();
+        CountlyConfig config = new CountlyConfig(ApplicationProvider.getApplicationContext(), "appkey", "http://test.count.ly").setDeviceId("1234").setLoggingEnabled(true).enableCrashReporting();
+        config.sdkInternalLimits.setMaxKeyLength(2);
+        mCountly.init(config);
+
+        Countly.sharedInstance().userProfile().setProperties(TestUtils.map(
+            ModuleUserProfile.BYEAR_KEY, 2000,
+            ModuleUserProfile.EMAIL_KEY, "email",
+            ModuleUserProfile.GENDER_KEY, "Male",
+            ModuleUserProfile.PHONE_KEY, "phone",
+            ModuleUserProfile.ORG_KEY, "org",
+            ModuleUserProfile.USERNAME_KEY, "username",
+            ModuleUserProfile.NAME_KEY, "name",
+            ModuleUserProfile.PICTURE_KEY, "picture",
+            "custom1", "value1",
+            "custom2", 23,
+            "hair", "black"
+        ));
+        Countly.sharedInstance().userProfile().save();
+
+        validateUserProfileRequest(mCountly, TestUtils.map(
+                ModuleUserProfile.BYEAR_KEY, 2000,
+                ModuleUserProfile.EMAIL_KEY, "email",
+                ModuleUserProfile.GENDER_KEY, "Male",
+                ModuleUserProfile.PHONE_KEY, "phone",
+                ModuleUserProfile.ORG_KEY, "org",
+                ModuleUserProfile.USERNAME_KEY, "username",
+                ModuleUserProfile.NAME_KEY, "name",
+                ModuleUserProfile.PICTURE_KEY, "picture"
+            ), TestUtils.map(
+                "cu", "23", // because in user profiles, all values are stored as strings
+                "ha", "black")
+        );
+    }
+
+    private void validateUserProfileRequest(Countly countly, Map<String, Object> predefined, Map<String, Object> custom) throws JSONException {
+        Map<String, String>[] RQ = TestUtils.getCurrentRQ();
+        Assert.assertEquals(1, RQ.length);
+        JSONObject userDetails = new JSONObject(RQ[0].get("user_details"));
+        Assert.assertEquals(userDetails.length(), predefined.size() + 1);
+        JSONObject customData = userDetails.getJSONObject("custom");
+        Assert.assertEquals(customData.length(), custom.size());
+        userDetails.remove("custom");
+        for (Map.Entry<String, Object> entry : predefined.entrySet()) {
+            Assert.assertEquals(entry.getValue(), userDetails.get(entry.getKey()));
+        }
+
+        for (Map.Entry<String, Object> entry : custom.entrySet()) {
+            Assert.assertEquals(entry.getValue(), customData.get(entry.getKey()));
+        }
     }
 }