+++ sdk/src/main/java/ly/count/android/sdk/ModuleUserProfile.java
@@ -224,6 +224,8 @@ public class ModuleUserProfile extends ModuleBase {
                 return;
             }
 
+            String truncatedKey = UtilsInternalLimits.truncateKeyLength(key, _cly.config_.sdkInternalLimits.maxKeyLength, _cly.L, "[ModuleUserProfile] modifyCustomData");
+
             if (customMods == null) {
                 customMods = new HashMap<>();
             }
@@ -232,14 +234,14 @@ public class ModuleUserProfile extends ModuleBase {
                 ob = new JSONObject();
                 ob.put(mod, value);
             } else {
-                if (customMods.containsKey(key)) {
-                    ob = customMods.get(key);
+                if (customMods.containsKey(truncatedKey)) {
+                    ob = customMods.get(truncatedKey);
                 } else {
                     ob = new JSONObject();
                 }
                 ob.accumulate(mod, value);
             }
-            customMods.put(key, ob);
+            customMods.put(truncatedKey, ob);
             isSynced = false;
         } catch (JSONException e) {
             e.printStackTrace();
@@ -277,7 +279,8 @@ public class ModuleUserProfile extends ModuleBase {
             }
 
             if (!isNamed) {
-                dataCustomFields.put(key, value.toString());
+                String truncatedKey = UtilsInternalLimits.truncateKeyLength(key, _cly.config_.sdkInternalLimits.maxKeyLength, _cly.L, "[ModuleUserProfile] setPropertiesInternal");
+                dataCustomFields.put(truncatedKey, value.toString());
             }
         }
 
@@ -288,6 +291,7 @@ public class ModuleUserProfile extends ModuleBase {
         if (custom == null) {
             custom = new HashMap<>();
         }
+
         custom.putAll(dataCustomFields);
 
         isSynced = false;